requirements.txt
streamlit
requests
gtts
google-genai